refactor(getApp): revert dispatch in useeffect
